refactor: 바뀐 훅에 따라서 재연결
