+++ src/components/CartIcon/CartIcon.tsx
@@ -1,10 +1,11 @@
 import CartIconImage from '@/assets/Cart.svg';
 import ImageBox from '@/components/common/ImageBox/ImageBox';
 import styles from './CartIcon.module.css';
-import { useCartItemListContext } from '@/hooks/useCartItemList';
+import useCartItemListContext from '@/hooks/useCartItemListContext';
 
 export default function CartIcon() {
   const { cartItemList } = useCartItemListContext();
+
   return (
     <div className={styles.container}>
       <ImageBox className={styles.cartIcon} src={CartIconImage} width={32} height={32} />
